requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pydantic
pydantic-settings
python-jose[cryptography]
argon2-cffi
groq
python-dotenv
python-multipart
email-validator
